Move color_tiles event handling to its dedicated file
