+++ src/gui/mod.rs
@@ -1,6 +1,6 @@
 pub mod builder;
+mod color_tiles;
 mod effect_browser_tile;
 pub mod enums;
-mod keyboard_color_tiles;
 mod menu_bar;
 mod options_tile;
